# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ecommerce API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update category (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── BANNERS ────────────────────────────────────────────────────────────────
  /banners:
    get:
      operationId: listBanners
      tags: [banners]
      summary: List active banners
      responses:
        "200":
          description: Banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
    post:
      operationId: createBanner
      tags: [banners]
      summary: Create banner (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
  /banners/{id}:
    patch:
      operationId: updateBanner
      tags: [banners]
      summary: Update banner (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
    delete:
      operationId: deleteBanner
      tags: [banners]
      summary: Delete banner (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── ORDERS ─────────────────────────────────────────────────────────────────
...
  /addresses:
    get:
      operationId: listAddresses
      tags: [addresses]
      summary: List user addresses
      responses:
        "200":
          description: Addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
    post:
      operationId: createAddress
      tags: [addresses]
      summary: Create address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
  /addresses/{id}:
    patch:
      operationId: updateAddress
      tags: [addresses]
      summary: Update address
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
    delete:
      operationId: deleteAddress
      tags: [addresses]
      summary: Delete address
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── WISHLIST ───────────────────────────────────────────────────────────────
...
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Get store settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
    patch:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Update store settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users (admin only)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        productCount:
          type: integer
    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        image:
          type: string
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        image:
          type: string

    # ─── BANNER ───────────────────────────────────────────────────────────────
    Banner:
      type: object
      required: [id, title, imageUrl, isActive]
      properties:
        id:
          type: integer
        title:
          type: string
        subtitle:
          type: ["string", "null"]
        imageUrl:
          type: string
        linkUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        sortOrder:
          type: integer
    BannerInput:
      type: object
      required: [title, imageUrl]
      properties:
        title:
          type: string
        subtitle:
          type: string
        imageUrl:
          type: string
        linkUrl:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
    BannerUpdate:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
        imageUrl:
          type: string
        linkUrl:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer

    # ─── ORDER ────────────────────────────────────────────────────────────────
...
    Address:
      type: object
      required: [id, userId, fullName, phone, addressLine, city, state, pincode]
      properties:
        id:
          type: integer
        userId:
          type: string
        fullName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        addressLine:
          type: string
        landmark:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        isDefault:
          type: boolean
    AddressInput:
      type: object
      required: [fullName, phone, addressLine, city, state, pincode]
      properties:
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        addressLine:
          type: string
        landmark:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        isDefault:
          type: boolean
    AddressUpdate:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        addressLine:
          type: string
        landmark:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        isDefault:
          type: boolean

    # ─── WISHLIST ─────────────────────────────────────────────────────────────
...
    AdminSettings:
      type: object
      required: [id, storeName]
      properties:
        id:
          type: integer
        storeName:
          type: string
        storeTagline:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        currency:
          type: string
        deliveryCharge:
          type: number
        freeDeliveryAbove:
          type: ["number", "null"]
        gstPercent:
          type: number
        whatsappNumber:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        adminClerkUserId:
          type: ["string", "null"]
        trustBadge1:
          type: ["string", "null"]
        trustBadge2:
          type: ["string", "null"]
        trustBadge3:
          type: ["string", "null"]
    AdminSettingsUpdate:
      type: object
      properties:
        storeName:
          type: string
        storeTagline:
          type: string
        logoUrl:
          type: string
        currency:
          type: string
        deliveryCharge:
          type: number
        freeDeliveryAbove:
          type: number
        gstPercent:
          type: number
        whatsappNumber:
          type: string
        instagramUrl:
          type: string
        adminClerkUserId:
          type: string
        trustBadge1:
          type: string
        trustBadge2:
          type: string
        trustBadge3:
          type: string
    AppUser:
      type: object
      required: [id, clerkUserId, email]
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        email:
          type: string
        fullName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        createdAt:
          type: string
    UserListResponse:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AppUser"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer