# lib/api-zod/src/generated/types/addressUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ecommerce API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AddressUpdate {
  fullName?: string;
  phone?: string;
  email?: string;
  addressLine?: string;
  landmark?: string;
  city?: string;
  state?: string;
  pincode?: string;
  isDefault?: boolean;
}