# lib/api-zod/src/generated/types/adminSettingsUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ecommerce API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminSettingsUpdate {
  storeName?: string;
  storeTagline?: string;
  logoUrl?: string;
  currency?: string;
  deliveryCharge?: number;
  freeDeliveryAbove?: number;
  gstPercent?: number;
  whatsappNumber?: string;
  instagramUrl?: string;
  adminClerkUserId?: string;
  trustBadge1?: string;
  trustBadge2?: string;
  trustBadge3?: string;
}